requirements.txt
playwright==1.47.2
requests>=2.32.3
beautifulsoup4>=4.12.3
openai>=1.47.0
tenacity>=8.3.0
pymupdf>=1.24.10
python-dotenv>=1.0.1
tqdm>=4.66.5
